Improvements to the Harmonic Balance engine
Added support to BJT devices.
Initial condition is now only the DC analysis result.
Added some code sections to use the secant method to calculate the
transistor conductances instead of using algebraic formulations. Still
need to understand how to place that in the conductance matrix since the
secant only worked for the diode.
Now the HB solution returns a complex voltage array with the spectra and
also a matrix with the time-domain waveforms for plotting.