Initial implementation of oscprobe for HB sim
Added the IdealHarmonicFilter component which is a short for the 1st
harmonic of HB and an open for all other frequencies including DC

Improved HB implementation details and added a plot function for
frequency and time-domain waveforms

Implemented 1st testbench of oscillator to compare with ADS simulation.
It converges with source stepping for a 2.3V input, which is a good
sign.

TODO: needed to make a weird modification to the Inductor component in
order to be used in HB simulation. This needs to improved.